+++ nvim/init.lua
@@ -1,4 +1,4 @@
-require('plugins')
+require('lazy')
 require('settings')
 require('mappings')
 require('theme')
